+++ libft/ft_strlen.c
@@ -3,18 +3,18 @@
 /*                                                        :::      ::::::::   */
 /*   ft_strlen.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: abouramt <marvin@42.fr>                    +#+  +:+       +#+        */
+/*   By: abouramt <abouramt@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/07 15:39:47 by abouramt          #+#    #+#             */
-/*   Updated: 2023/11/07 15:40:14 by abouramt         ###   ########.fr       */
+/*   Updated: 2024/05/22 18:46:26 by abouramt         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-size_t	ft_strlen(const char *c)
+int	ft_strlen(const char *c)
 {
-	size_t	i;
+	int	i;
 
 	i = 0;
 	while (c[i])
